+++ aioworkers/core/plugin.py
@@ -1,41 +1,48 @@
+import functools
 import logging
 import sys
 from pathlib import Path
+from typing import Dict, Iterable, Optional
 
 from . import config, formatter
 
 logger = logging.getLogger(__name__)
 
 
-def load_plugin(module: str, force=False):
-    if force:
+def load_plugin(module: str,
+                force: bool = False,
+                *,
+                cache: Dict = {}) -> Optional['Plugin']:
+    if module in cache:
+        return cache[module]
+    elif force:
         pass
     elif module in sys.modules:
-        return
+        return None
     try:
         m = __import__(module, fromlist=['plugin'])
     except ImportError:
-        return
-    if hasattr(m, 'plugin'):
-        plugin = m.plugin
-    else:
-        plugin = m
+        return None
+    plugin = getattr(m, 'plugin', m)
     if callable(plugin):
         plugin = plugin()
-    if isinstance(plugin, Plugin):
-        return plugin
-    return ProxyPlugin(plugin)
+    if not isinstance(plugin, Plugin):
+        plugin = ProxyPlugin(plugin)
+    cache[module] = plugin
+    return plugin
 
 
-def get_names():
-    for path in sys.path:
-        path = Path(path)
+@functools.lru_cache(None)
+def get_names() -> Iterable[str]:
+    for s in sys.path:
+        path = Path(s)
         if not path.is_dir():
             continue
         for i in path.glob('aioworkers*'):
             yield i.with_suffix('').name
 
 
+@functools.lru_cache(None)
 def search_plugins(*modules, force=False):
     result = []
     if not modules:
@@ -65,13 +72,17 @@ class Plugin:
     def add_arguments(self, parser):
         pass
 
+    def parse_known_args(self, args, namespace):
+        """argparse method"""
+        return namespace, args
+
 
 class ProxyPlugin(Plugin):
     def __init__(self, original):
         self._original = original
         for i in (
-                'formatters', 'config_loaders', 'configs',
-                'get_config', 'add_arguments',
+            'formatters', 'config_loaders', 'configs',
+            'get_config', 'add_arguments',
         ):
             v = getattr(original, i, None)
             if v: