+++ tests/test_config.py
@@ -5,8 +5,9 @@ from pathlib import Path
 import pytest
 
 from aioworkers import utils
-from aioworkers.core.config import (Config, IniLoader, MergeDict,
-                                    StringReplaceLoader, ValueExtractor)
+from aioworkers.core.config import (
+    Config, IniLoader, MergeDict, StringReplaceLoader, ValueExtractor
+)
 
 
 def test_dict_create():
@@ -122,10 +123,19 @@ def test_ini():
         list2:[1,2]
         list3=[1,2]
         list4 = [1,2]
+        bool1: true
+        bool2: false
+        bool3: True
+        bool4: False
+        bool5: On
+        bool6: Off
+        int7: 1
         """)
     for k, v in d['sec'].items():
         if k.startswith('int'):
             assert isinstance(v, int)
+        elif k.startswith('bool'):
+            assert isinstance(v, bool)
         elif k.startswith('float'):
             assert isinstance(v, float)
         elif k.startswith('list'):