+++ tests/test_net_web.py
@@ -2,16 +2,35 @@ import pytest
 
 
 @pytest.fixture
+def aioworkers(aioworkers):
+    aioworkers.plugins.append('aioworkers.net.web')
+    return aioworkers
+
+
+@pytest.fixture
+def config_yaml():
+    return """
+    app.resources:
+        /api:
+            get: .data
+        /api/str:
+            get: .str_data
+        /api/bin:
+            get: .bin_data
+    data: 1
+    str_data: asdf
+    storage.cls: aioworkers.storage.http.Storage
+    """
+
+
+@pytest.fixture
 def config(config):
-    config.load_plugins('aioworkers.net.web')
-    config.update({
-        'app.resources': {'/api': {'get': '.data'}},
-        'data': 1,
-        'storage.cls': 'aioworkers.storage.http.Storage',
-    })
+    config.update(bin_data=b'qwerty')
     return config
 
 
 async def test_web_server(context):
     url = context.http.url
     assert 1 == await context.storage.get(url / 'api')
+    assert b'asdf' == await context.storage.get(url / 'api/str')
+    assert b'qwerty' == await context.storage.get(url / 'api/bin')